requirements.txt
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-allauth==0.60.1
djangorestframework-simplejwt == 5.4.0
django-countries==7.5.1
django-filter==23.5
Pillow==10.2.0
python-decouple==3.8
stripe==7.11.0
psycopg2-binary==2.9.9
dj-rest-auth==5.0.2
python-magic==0.4.27